Move planet shaders to separate files & load them directly in js, instead of using html elements
