+++ orbitez/services/planet/infra/createProgram.js
@@ -1,0 +1,14 @@
+function createProgram(gl, vertexShader, fragmentShader) {
+    const program = gl.createProgram();
+    gl.attachShader(program, vertexShader);
+    gl.attachShader(program, fragmentShader);
+    gl.linkProgram(program);
+    const success = gl.getProgramParameter(program, gl.LINK_STATUS);
+    if (success) {
+        return program;
+    }
+
+    gl.deleteProgram(program);
+};
+
+export default createProgram;
