+++ orbitez/services/planet/infra/createShader.js
@@ -1,0 +1,13 @@
+function createShader(gl, type, source) {
+    const shader = gl.createShader(type);
+    gl.shaderSource(shader, source);
+    gl.compileShader(shader);
+    const success = gl.getShaderParameter(shader, gl.COMPILE_STATUS);
+    if (success) {
+        return shader;
+    }
+
+    gl.deleteShader(shader);
+};
+
+export default createShader;
